ref(metrics): Remove legacy conditional tagging code
